CollectImport: import survey description(s)
